refactor: rename CheckBulkPermissionAsync to CheckBulkPermissionsAsync
Renamed all instances of the method `CheckBulkPermissionAsync` to `CheckBulkPermissionsAsync` in both ISpiceDbClient and SpiceDbClient classes. This change improves readability and consistency across the codebase.
refactor: rename CheckBulkPermissionAsync to CheckBulkPermissionsAsync

Renamed all instances of the method `CheckBulkPermissionAsync` to `CheckBulkPermissionsAsync` in both ISpiceDbClient and SpiceDbClient classes. This change improves readability and consistency across the codebase. Also, updated the version number in SpiceDb.csproj from 1.4.6 to 1.4.7 reflecting these changes.