feat: Add SpiceWeaver package and update schema handling
Added the SpiceWeaver package to both the Example and SchemaParsingTest projects. This new dependency will enhance our ability to handle schemas. Also, updated how we deal with 'schema.zedj' in the SchemaParsingTest project by adding it as an AdditionalFile with specific attributes for better management. Lastly, a reference to Administrator relation was added in Program.cs file of SchemaParsingTest project for testing purposes.